fixed tests, moved util to component
