+++ src/react/src/__tests__/utils.tests.js
@@ -5,9 +5,6 @@ const mapValues = require('lodash/mapValues');
 const isEqual = require('lodash/isEqual');
 const includes = require('lodash/includes');
 const turf = require('@turf/turf');
-const React =  require('react');
-const { render } =  require('@testing-library/react');
-const { BrowserRouter } =  require('react-router-dom');
 
 const {
     makeFacilityListsURL,
@@ -103,10 +100,6 @@ const {
     CLAIM_A_FACILITY,
     componentsWithErrorMessage,
 } = require('../util/constants');
-const {
-    LoadingIndicator,
-    AuthNotice,
-    ErrorsList,} = require('../util/checkComponentStatus').default;
 
 const COLOURS = require('../util/COLOURS').default;
 
@@ -2248,36 +2241,3 @@ describe('getSelectStyles', () => {
         expect(controlStyles['&:hover']).toEqual({ borderColor: false });
     });
 });
-
-describe('CheckComponentStatus components', () => {
-    const expectedTitle = 'Expected title';
-
-      it('renders LoadingIndicator with title', () => {
-          const { getByText } = render(<LoadingIndicator title={expectedTitle} />);
-
-          expect(getByText(expectedTitle)).toBeInTheDocument();
-      });
-
-      it('renders AuthNotice with title and text', () => {
-        const text = 'Sign in to view your Open Supply Hub facility claims.';
-        const { getByText } = render(
-            <BrowserRouter>
-               <AuthNotice title={expectedTitle} />
-            </BrowserRouter>
-        )
-        expect(getByText(expectedTitle)).toBeInTheDocument();
-        expect(getByText(text)).toBeInTheDocument();
-      });
-
-      it('renders ErrorsList with title and errors', () => {
-        const errors =  [
-         'This field is required!','The right must be greater than left.',
-        ] ;
-        const { getByText } = render(<ErrorsList title={expectedTitle} errors={errors}/>);
-
-        expect(getByText(expectedTitle)).toBeInTheDocument();
-        expect(getByText(errors[0])).toBeInTheDocument();
-        expect(getByText(errors[1])).toBeInTheDocument();
-      });
-
-  });
